Update to fix issue #4
Seems as though if the BLE device on the pi disconnects before discovery, listeners are left. These become active later on when an actual connect to the device is successful and fires at the same time. This can cause a memory leak as well as issues where the system does not get a proper read from the device and the user sees the decide as non responsive in home kit.